Added comments, explanations and information to ConnectionManager
Added comments to the file to help understand the uses and reasoning behind what was done with the file.

Moved around some lines of code to better fit where they're being used, changed getRequest to return a Weather object for now.

Commented out the getJSONArray method as it could be utilised better elsewhere. Please feel free to use the code.